Restructure python stuff
