+++ python/ocr/main.py
@@ -1,0 +1,5 @@
+def main():
+	print("Woah")
+
+if __name__ == "__main__":
+	main()
