+++ python/scraper/requirements.txt
@@ -1,8 +1,4 @@
 requests
 beautifulsoup4
-waitress
 colorama
-yapf
-fastapi
-uvicorn
 redis
